# packages/shared-types/openapi.yaml
openapi: 3.1.0
info:
  title: Snake Backlink Forge API
  version: 0.1.0
  description: SBF — SEO backlink automation SaaS API. Phase 2 endpoints + Phase 3 web auth.
servers:
  - url: https://snake-backlink-api.fly.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: sbf_live_*
      description: API key issued via Telegram bot. Format sbf_live_<random>.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthVerifyRequest:
      type: object
      required: [key]
      properties:
        key:
          type: string
          pattern: "^sbf_live_"
    AuthVerifyResponse:
      type: object
      required: [ok, user]
      properties:
        ok:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      required: [id, key_prefix]
      properties:
        id:
          type: string
          format: uuid
        key_prefix:
          type: string
...
paths:
...
  /api/v1/auth/verify:
    post:
      summary: Verify API key
      operationId: postApiV1AuthVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyRequest"
      responses:
        "200":
          description: API key is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthVerifyResponse"
        "401":
          description: Invalid key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Account banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...